refactor: move fluid classes from common to domain and rename and group
