+++ src/ecalc_neqsim_wrapper/mappings.py
@@ -2,10 +2,10 @@ from __future__ import annotations
 
 from pydantic import Field
 
+from libecalc.common.ecalc_base_model import EcalcBaseModel
 from libecalc.common.errors.exceptions import EcalcError
-from libecalc.common.fluid import EcalcBaseModel
 from libecalc.common.logger import logger
-from libecalc.domain.process.value_objects.fluid_stream.fluid_composition import FluidComposition
+from libecalc.domain.process.value_objects.fluid_stream.fluid_model import FluidComposition
 
 
 class NeqsimComposition(EcalcBaseModel):
