+++ src/libecalc/common/fluid.py
@@ -1,11 +1,11 @@
 from __future__ import annotations
 
+from enum import Enum
+
 from pydantic import BaseModel, ConfigDict, model_validator
 
-from libecalc.common.fluid_stream_type import FluidStreamType
 from libecalc.common.string.string_utils import to_camel_case
-from libecalc.domain.process.value_objects.fluid_stream.eos_model import EoSModel
-from libecalc.domain.process.value_objects.fluid_stream.fluid_composition import FluidComposition
+from libecalc.domain.process.value_objects.fluid_stream.fluid_model import FluidModel
 
 
 class EcalcBaseModel(BaseModel):
@@ -16,9 +16,9 @@ class EcalcBaseModel(BaseModel):
     )
 
 
-class FluidModel(EcalcBaseModel):
-    eos_model: EoSModel
-    composition: FluidComposition
+class FluidStreamType(str, Enum):
+    INGOING = "INGOING"
+    OUTGOING = "OUTGOING"
 
 
 class MultipleStreamsAndPressureStream(EcalcBaseModel):
