--- src/libecalc/common/fluid_stream_type.py
@@ -1,6 +1,0 @@
-from enum import Enum
-
-
-class FluidStreamType(str, Enum):
-    INGOING = "INGOING"
-    OUTGOING = "OUTGOING"
