+++ src/libecalc/domain/process/compressor/core/results.py
@@ -12,7 +12,7 @@ from libecalc.domain.process.core.results.compressor import (
 )
 from libecalc.domain.process.value_objects.chart.chart_area_flag import ChartAreaFlag
 from libecalc.domain.process.value_objects.fluid_stream import FluidStream
-from libecalc.domain.process.value_objects.fluid_stream.fluid_composition import FluidComposition
+from libecalc.domain.process.value_objects.fluid_stream.fluid_model import FluidComposition
 
 
 class CompressorTrainStageResultSingleTimeStep:
