--- src/libecalc/domain/process/value_objects/fluid_stream/eos_model.py
@@ -1,8 +1,0 @@
-from enum import Enum
-
-
-class EoSModel(str, Enum):
-    SRK = "SRK"
-    PR = "PR"
-    GERG_SRK = "GERG_SRK"
-    GERG_PR = "GERG_PR"
