+++ src/libecalc/domain/process/value_objects/fluid_stream/fluid_stream.py
@@ -4,8 +4,8 @@ from dataclasses import dataclass
 from functools import cached_property
 
 from libecalc.common.units import UnitConstants
-from libecalc.domain.process.value_objects.fluid_stream.conditions import ProcessConditions
 from libecalc.domain.process.value_objects.fluid_stream.exceptions import NegativeMassRateException
+from libecalc.domain.process.value_objects.fluid_stream.process_conditions import ProcessConditions
 from libecalc.domain.process.value_objects.fluid_stream.thermo_system import ThermoSystemInterface
