+++ src/libecalc/domain/process/value_objects/fluid_stream/process_conditions.py
