+++ src/libecalc/domain/process/value_objects/fluid_stream/thermo_system.py
@@ -3,9 +3,8 @@ from __future__ import annotations
 from functools import cached_property
 from typing import Protocol
 
-from libecalc.domain.process.value_objects.fluid_stream.conditions import ProcessConditions
-from libecalc.domain.process.value_objects.fluid_stream.eos_model import EoSModel
-from libecalc.domain.process.value_objects.fluid_stream.fluid_composition import FluidComposition
+from libecalc.domain.process.value_objects.fluid_stream.fluid_model import EoSModel, FluidComposition
+from libecalc.domain.process.value_objects.fluid_stream.process_conditions import ProcessConditions
 
 
 class ThermoSystemInterface(Protocol):
