+++ tests/ecalc_neqsim_wrapper/unit_tests/test_neqsim_fluid.py
@@ -5,8 +5,7 @@ import pytest
 
 from ecalc_neqsim_wrapper.exceptions import NeqsimComponentError
 from ecalc_neqsim_wrapper.thermo import NeqsimFluid, mix_neqsim_streams
-from libecalc.domain.process.value_objects.fluid_stream.eos_model import EoSModel
-from libecalc.domain.process.value_objects.fluid_stream.fluid_composition import FluidComposition
+from libecalc.domain.process.value_objects.fluid_stream.fluid_model import EoSModel, FluidComposition
 from tests.ecalc_neqsim_wrapper.conftest import MEDIUM_MW_19P4_COMPOSITION
