+++ tests/libecalc/domain/process/fluid_stream/test_thermo_constants.py
@@ -1,5 +1,5 @@
 from libecalc.domain.process.value_objects.fluid_stream.constants import ThermodynamicConstants
-from libecalc.domain.process.value_objects.fluid_stream.fluid_composition import FluidComposition
+from libecalc.domain.process.value_objects.fluid_stream.fluid_model import FluidComposition
 
 
 def validate_components():
